Index-functions
Several functions will be added that access tags by index. Finding a tag works now by finding the tagView of the tag. This makes it hard to keep data and tags in sync. And it exposes the tagViews themselves to the enduser. This might lead to inconsistencies between the users-code, TagListView and TagView.